+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - fixed a bug with the todo list calendars if the server url had a path in it (a big thank you to Dirk for the report)
 - more information was added to the debug information and you can now store it to a markdown file to help us finding problems
 - pressing return in the summary text line of a todo item now saves it
+- added basic markdown highlighting to the description edit field of the todo items
 
 ## 0.72
 - fixed a crash that occurred when all notes were removed